#25 Adding rich text editor widget TinyMCE, adding model and admin for presentations.
Building out the presentations add on for issue #25. Currently using the package django-tinymce (http://django-tinymce.readthedocs.org/en/latest/index.html) to allow for rich text editor widgets, specifically for the description field for Presentations which was described as supporting HTML -- the supporting static js files are included in this commit as well as package dependent urls and settings. Extending admin interface for presentations; not sure if this will be needed or will be needed for events as well. New migration to extend dcpython.events.models with new Presentation class.